convertNumber function was moved to FormatNumbers.kt
